Combine file source controllers
